[Cleanup] Extract shared timestep_embedding function
Move the duplicated timestep_embedding() function (GLIDE/DiT cos-then-sin
convention) into the shared timestep_embedding module.

Replaces 4 identical copies:
- bagel/bagel_transformer.py (staticmethod on TimestepEmbedder)
- nextstep_1_1/modeling_nextstep_heads.py (staticmethod on TimestepEmbedder)
- z_image/z_image_transformer.py (staticmethod on TimestepEmbedder)
- hunyuan_image3/hunyuan_image3_transformer.py (module-level function)
