+++ vllm_omni/diffusion/models/nextstep_1_1/modeling_nextstep_heads.py
@@ -6,10 +6,10 @@
 
 from __future__ import annotations
 
-import math
-
 import torch
 import torch.nn as nn
+
+from vllm_omni.model_executor.layers.timestep_embedding import timestep_embedding
 
 # ---------------------------------------------------------------------------
 # Utilities (inlined from remote utils/model_utils.py)
@@ -109,20 +109,8 @@ class TimestepEmbedder(nn.Module):
         )
         self.frequency_embedding_size = frequency_embedding_size
 
-    @staticmethod
-    def timestep_embedding(t: torch.Tensor, dim: int, max_period: float = 10000.0) -> torch.Tensor:
-        half = dim // 2
-        freqs = torch.exp(-math.log(max_period) * torch.arange(start=0, end=half, dtype=torch.float32) / half).to(
-            device=t.device
-        )
-        args = t[:, None].float() * freqs[None]
-        embedding = torch.cat([torch.cos(args), torch.sin(args)], dim=-1)
-        if dim % 2:
-            embedding = torch.cat([embedding, torch.zeros_like(embedding[:, :1])], dim=-1)
-        return embedding
-
     def forward(self, t: torch.Tensor) -> torch.Tensor:
-        t_freq = self.timestep_embedding(t, self.frequency_embedding_size)
+        t_freq = timestep_embedding(t, self.frequency_embedding_size)
         return self.mlp(t_freq.to(self.mlp[0].weight.dtype))
